+++ src/draw_best_line.c
@@ -6,7 +6,7 @@
 /*   By: lorey <loic.rey.vs@gmail.com>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/05 14:48:06 by lorey             #+#    #+#             */
-/*   Updated: 2025/04/06 02:56:36 by lorey            ###   LAUSANNE.ch       */
+/*   Updated: 2025/04/06 16:40:17 by lorey            ###   LAUSANNE.ch       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -128,7 +128,7 @@ void	draw_best_line(t_mlx_data *data)
 	ver_no = ver_best_line(data, &ver);
 	hor_no = hor_best_line(data, &hor);
 	data->ray_hor = sqrt(pow(hor.offset, 2) + pow(hor.delta_y - (double)data->player_y, 2));
-	data->ray_ver = sqrt(pow(ver.delta_x -data->player_x, 2) + pow(ver.offset, 2));
+	data->ray_ver = sqrt(pow(ver.delta_x - (double)data->player_x, 2) + pow(ver.offset, 2));
 	if (ver_no == -1)
 		draw_horizontal_line(data, &hor);
 		//draw_line(data, data->player_x, data->player_y, (int)(data->player_x + hor.offset), hor.delta_y, WHITE);
